cleanup: reorganized onboarding and profile file structure
