+++ components/onboarding/Onboarding.tsx
@@ -14,10 +14,10 @@ import {
   OnboardingSchema,
 } from "@/lib/validations/UserSchema";
 
-import BasicInformation from "@/components/onboarding/BasicInformation";
-import LearningGoals from "@/components/onboarding/LearningGoals";
-import KnowledgeLevel from "@/components/onboarding/KnowledgeLevel";
-import Availability from "@/components/onboarding/Availability";
+import BasicInformation from "@/components/profile/BasicInformation";
+import LearningGoals from "@/components/profile/LearningGoals";
+import KnowledgeLevel from "@/components/profile/KnowledgeLevel";
+import Availability from "@/components/profile/Availability";
 import OnboardingVisualStepper from "@/components/onboarding/OnboardingVisualStepper";
 import Button from "@/components/shared/ui/Button";
